requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.6
docutils==0.22
filelock==3.18.0
HTMLParser==0.0.2
huggingface_hub==0.34.4
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
keyring==25.6.0
matplotlib==3.10.5
mock==5.2.0
Pillow==11.3.0
playwright==1.54.0
protobuf==6.31.1
pydantic==2.11.7
pyOpenSSL==25.1.0
python-dotenv==1.1.1
railroad==0.5.0
redis==6.4.0
seaborn==0.13.2
Sphinx==8.2.3
thread==2.0.5
torch==2.8.0+cu126
tornado==6.5.2
transformers==4.55.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
